+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Note - after update to 3.15.x need remove the accessory frome Home app and add it again.
 
+## [3.22.0] - (07.02.2023)
+## Changes
+- added possibility load inputs from device (release)
+- removed duplicated display type *sensors* from inputs section
+- config.schema updated
+- bump dependencies
+- cleanup
+
 ## [3.21.0] - (25.01.2023)
 ## Changes
 - added possibility load inputs and quick selects from device (test phase)